Client/requirements.txt
#os
pydantic==1.9.0
requests==2.27.1
python-dotenv==0.20.0
